+++ TeamProject/api/post.py
@@ -84,7 +84,7 @@ def bestpost_all():
 		returnlist.append(post.serialize)
 		returnlist[i].update(board_name=post.board.board_name)		# board_name = 해당 글이 속하는 게시판 이름
 	return jsonify(returnlist), 200      # json으로 게시글 목록 리턴
-
+ 
 ### 해당 게시판 베스트 게시글 ###
 @api.route('/bestpost/<id>', methods=['GET'])			# 베스트 게시글
 def bestpost_board(id):
@@ -122,8 +122,13 @@ def post_get():
 
 ### 게시글 (글쓰기) ###
 @api.route('/post', methods=['POST'])
-# @jwt_required
+@jwt_required
 def post_post():
+	check_user= get_jwt_identity()
+	if check_user != 'GM':
+		access_user = User.query.filter(User.userid == check_user).first()			# 꺼낸 토큰이 유효한 토큰인지 확인
+		if access_user is None:
+			return {'error':'잘못된 토큰입니다.'}, 403			# 1아니면 0 값을 보내야하는데 다른 값을 보내는 경우 오류
 	# POST
 	if request.method == 'POST':
 		data = request.get_json()
@@ -134,15 +139,16 @@ def post_post():
 		board_name = data.get('board_name')			# 해당하는 게시판의 이름
 
 		# 블랙리스트 확인
-		user = User.query.filter(User.id == userid).first()
-		if user.Black_set_user:
-			black = Blacklist.query.filter(Blacklist.userid == userid).first()
-			if black.punishment_end > datetime.now():
-				return jsonify({'error':'현재 당신의 아이디는 게시글을 쓸 수 없습니다.'}), 403
-			else :		# 블랙은 되었었으나, 정지가 풀리는 날 이후인 경우 블랙리스트에서 제외
-				# black = Blacklist.query.filter(Blacklist.userid == userid).first()
-				db.session.delete(black)
-				db.session.commit()
+		if check_user != 'GM':
+			user = User.query.filter(User.id == userid).first()
+			if user.Black_set_user:
+				black = Blacklist.query.filter(Blacklist.userid == userid).first()
+				if black.punishment_end > datetime.now():
+					return jsonify({'error':'현재 당신의 아이디는 게시글을 쓸 수 없습니다.'}), 403
+				else :		# 블랙은 되었었으나, 정지가 풀리는 날 이후인 경우 블랙리스트에서 제외
+					# black = Blacklist.query.filter(Blacklist.userid == userid).first()
+					db.session.delete(black)
+					db.session.commit()
 
 		if not subject:
 			return jsonify({'error': '제목이 없습니다.'}), 400
@@ -182,7 +188,12 @@ def post_detail(id):
 @api.route('/post/<id>', methods=['PUT', 'DELETE'])
 @jwt_required
 def post_detail_modified(id):
-
+	check_user= get_jwt_identity()
+	if check_user != 'GM':
+		access_user = User.query.filter(User.userid == check_user).first()			# 꺼낸 토큰이 유효한 토큰인지 확인
+		if access_user is None:
+			return {'error':'잘못된 토큰입니다.'}, 403			# 1아니면 0 값을 보내야하는데 다른 값을 보내는 경우 오류
+	
 	# DELETE
 	if request.method == 'DELETE':                            # 삭제
 		post = Post.query.filter(Post.id == id).first()
@@ -234,6 +245,12 @@ def comment(id):
 @api.route('/comment/<id>',methods=['PUT', 'POST', 'DELETE'])		# id = post의 id
 @jwt_required
 def comment_modified(id):
+	check_user= get_jwt_identity()
+	if check_user != 'GM':
+		access_user = User.query.filter(User.userid == check_user).first()			# 꺼낸 토큰이 유효한 토큰인지 확인
+		if access_user is None:
+			return {'error':'잘못된 토큰입니다.'}, 403			# 1아니면 0 값을 보내야하는데 다른 값을 보내는 경우 오류
+	
 	# POST
 	if request.method == 'POST':
 		data = request.get_json()
@@ -242,15 +259,16 @@ def comment_modified(id):
 		create_date = datetime.now()
 
 		# 블랙리스트 확인
-		user = User.query.filter(User.id == userid).first()
-		if user.Black_set_user:
-			black = Blacklist.query.filter(Blacklist.userid == userid).first()
-			if black.punishment_end > datetime.now():
-				return jsonify({'error':'현재 당신의 아이디는 댓글을 쓸 수 없습니다.'}),201
-			else :		# 블랙은 되었었으나, 정지가 풀리는 날 이후인 경우 블랙리스트에서 제외
-				# black = Blacklist.query.filter(Blacklist.userid == userid).first()
-				db.session.delete(black)
-				db.session.commit()
+		if check_user != 'GM':
+			user = User.query.filter(User.id == userid).first()
+			if user.Black_set_user:
+				black = Blacklist.query.filter(Blacklist.userid == userid).first()
+				if black.punishment_end > datetime.now():
+					return jsonify({'error':'현재 당신의 아이디는 댓글을 쓸 수 없습니다.'}),201
+				else :		# 블랙은 되었었으나, 정지가 풀리는 날 이후인 경우 블랙리스트에서 제외
+					# black = Blacklist.query.filter(Blacklist.userid == userid).first()
+					db.session.delete(black)
+					db.session.commit()
 
 		if not content:
 			return jsonify({'error': '내용이 없습니다.'}), 400
@@ -302,8 +320,7 @@ def comment_modified(id):
 def postlike(id):
 	user_id = get_jwt_identity()
 	access_user = User.query.filter(User.userid == user_id).first()
-
-	if access_user is None:
+	if access_user is None and user_id !='GM':
 		print("None")
 		g.user = None
 	else:
@@ -332,7 +349,7 @@ def commentlike(id):
 	user_id = get_jwt_identity()
 	access_user = User.query.filter(User.userid == user_id).first()
 
-	if access_user is None:
+	if access_user is None and user_id != 'GM':
 		print("None")
 		g.user = None
 	else:
@@ -430,7 +447,7 @@ def post_uploadimg(id):
 def report_post(id):
 	userid = get_jwt_identity()
 	access_user = User.query.filter(User.userid == userid).first()
-	if access_user is None:		# 유효하지 않은 토큰이 들어있는 경우
+	if access_user is None and userid != 'GM':		# 유효하지 않은 토큰이 들어있는 경우
 		return jsonify({'error':'Bad Access Token'}), 403
 
 	g.user = access_user
@@ -449,7 +466,7 @@ def report_post(id):
 def report_comment(id):
 	userid = get_jwt_identity()
 	access_user = User.query.filter(User.userid == userid).first()
-	if access_user is None:		# 유효하지 않은 토큰이 들어있는 경우
+	if access_user is None and userid != 'GM':		# 유효하지 않은 토큰이 들어있는 경우
 		return jsonify({'error':'Bad Access Token'}), 403
 
 	g.user = access_user
